# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Profile {
  name: string;
  affiliateId: string;
  packageLabel: string;
  photoUrl: string | null;
}

export interface Earnings {
  today: number;
  sevenDay: number;
  thirtyDay: number;
  allTime: number;
}

export interface DashboardResponse {
  profile: Profile;
  earnings: Earnings;
}

export interface UpdateProfileRequest {
  name: string;
  affiliateId: string;
  packageLabel: string;
  photoUrl?: string | null;
}

export interface UpdateEarningsRequest {
  /** @minimum 0 */
  amount: number;
}