# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: dashboard
    description: Public affiliate dashboard data
  - name: admin
    description: Admin controls for profile and earnings
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Get dashboard data
      description: Returns the affiliate profile and computed earning totals
      responses:
        "200":
          description: Dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardResponse"
  /admin/profile:
    put:
      operationId: updateProfile
      tags: [admin]
      summary: Update affiliate profile
      description: Updates the displayed name, affiliate ID, package label, and photo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardResponse"
  /admin/earnings:
    put:
      operationId: updateTodayEarning
      tags: [admin]
      summary: Set today's earning amount
      description: Sets an absolute amount for today, which automatically recomputes 7-day, 30-day, and all-time rolling totals
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEarningsRequest"
      responses:
        "200":
          description: Updated dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Profile:
      type: object
      properties:
        name:
          type: string
        affiliateId:
          type: string
        packageLabel:
          type: string
        photoUrl:
          type: string
          nullable: true
      required:
        - name
        - affiliateId
        - packageLabel
        - photoUrl
    Earnings:
      type: object
      properties:
        today:
          type: number
        sevenDay:
          type: number
        thirtyDay:
          type: number
        allTime:
          type: number
      required:
        - today
        - sevenDay
        - thirtyDay
        - allTime
    DashboardResponse:
      type: object
      properties:
        profile:
          $ref: "#/components/schemas/Profile"
        earnings:
          $ref: "#/components/schemas/Earnings"
      required:
        - profile
        - earnings
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
        affiliateId:
          type: string
        packageLabel:
          type: string
        photoUrl:
          type: string
          nullable: true
      required:
        - name
        - affiliateId
        - packageLabel
    UpdateEarningsRequest:
      type: object
      properties:
        amount:
          type: number
          minimum: 0
      required:
        - amount