# lib/api-zod/src/generated/types/dashboardResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Earnings } from './earnings';
import type { Profile } from './profile';

export interface DashboardResponse {
  profile: Profile;
  earnings: Earnings;
}