# lib/api-zod/src/generated/types/earnings.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Earnings {
  today: number;
  sevenDay: number;
  thirtyDay: number;
  allTime: number;
}