# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  name: string;
  affiliateId: string;
  packageLabel: string;
  photoUrl: string | null;
}